Move tests from resp2 to bytesutil package
